more formatting

+++ helpers/AppHelper.sh
@@ -36,11 +36,11 @@ Choose an option:  "
     case $ans in
     1)
         echo "Start Portainer"
-	    sudo docker container start portainer &1>/dev/null && $CALLBIGGAP && echo "Portainer should now be available on localhost:9000"
+	    $CALLBIGGAP && sudo docker container start portainer &1>/dev/null && $CALLBIGGAP && echo "Portainer should now be available on localhost:9000"
         portaineroptions
         ;;
     2)
-        sudo docker container stop portainer &1>/dev/null && $CALLBIGGAP && echo "Portainer has been shut down"
+        $CALLBIGGAP && sudo docker container stop portainer &1>/dev/null && $CALLBIGGAP && echo "Portainer has been shut down"
         portaineroptions
         ;;
     3)
